--- .env
@@ -1,1 +1,0 @@
-RECOIL_DUPLICATE_ATOM_KEY_CHECKING_ENABLED=false
