+++ apis/config/index.ts
@@ -14,10 +14,14 @@ axiosInstance.interceptors.request.use(
     let token = null;
 
     /* refreshToken 토큰 사용 api 주소  */
-    if (config.url === REFRESH_URL) {
-      token = localStorage.getItem('refreshToken');
-    } else {
-      token = localStorage.getItem('accessToken');
+
+    // Server-side 에서는 window가 없기 때문에 local storage 사용 불가
+    if (typeof window !== 'undefined') {
+      if (config.url === REFRESH_URL) {
+        token = localStorage.getItem('refreshToken');
+      } else {
+        token = localStorage.getItem('accessToken');
+      }
     }
 
     if (token !== null) {
@@ -35,33 +39,27 @@ axiosInstance.interceptors.request.use(
 // 응답 인터셉터 설정
 axiosInstance.interceptors.response.use(
   (response) => {
+
     // 응답 데이터 전에 수행할 작업
 
     return response;
   },
   async (error) => {
-    // 응답 에러 처리
-    const {
-      config,
-      response: { status, data },
-    } = error;
+    const { config, response } = error;
 
-    // if (data.code === 401) {
-    //   const token = localStorage.getItem('refreshToken');
-    //   await postRefreshToken(token as string);
+    // Server-side에서 response 객체가 존재하지 않을 때 에러가 발생하기 때문에, response가 존재하면 코드 실행하는 조건 추가
+    if (response && response.status && response.data) {
+      const { status, data } = response;
+      toast.error(data?.message);
+      console.log(data);
+    }
 
-    //   return axios(config);
-    // }
-
-    toast.error(data?.message);
-    console.log(data);
     return Promise.reject(error);
   },
 );
 
 export default axiosInstance;
 
+const BASE_URL = 'http://43.201.195.195:8080';
 
-const BASE_URL = 'http://43.201.195.195:8080'
-
-export const ec2 = axios.create({ baseURL: BASE_URL })
+export const ec2 = axios.create({ baseURL: BASE_URL });
