+++ components/customer/Notice.tsx
@@ -1,36 +1,24 @@
-import React, { FC, useState, Fragment, useMemo } from "react";
+import React, { FC, Fragment, useMemo } from "react";
 import NoticeList from "./NoticeList";
 import { useInfiniteQuery } from "@tanstack/react-query";
-import { atom, useRecoilValue } from "recoil";
 import { useInView } from 'react-intersection-observer'
-import axios from "axios";
-import { lastArticleIdState, sizeState } from "@/atoms/noticeAtom";
 import { NoticePostType } from "@/utils/types/responseType";
+import axiosInstance from "@/apis/config";
 
 const Notice: FC<{}> = () => {
-    const lastArticleId = useRecoilValue(lastArticleIdState);
-    const size = useRecoilValue(sizeState);
-    const fetchPreviousPageAsync = async () => {
-        await fetchPreviousPage();
-    };
-
-    const fetchNextPageAsync = async () => {
-        await fetchNextPage();
-    };
 
     const { ref, inView } = useInView()
 
-    const path = "http://43.201.195.195:8080";
-    const { data, isFetchingNextPage, isFetchingPreviousPage, fetchNextPage, fetchPreviousPage, hasPreviousPage } =
+    const { data, fetchNextPage } =
         useInfiniteQuery<NoticePostType[], Error>(
             ["notices"],
             async ({ pageParam = 0 }) => {
                 if (!pageParam) {
-                    const res = await axios.get(`${path}/api/notices?size=10&firstPage=true`);
+                    const res = await axiosInstance.get(`/api/notices?size=10&firstPage=true`);
                     return res.data;
                 }
 
-                const res = await axios.get(`${path}/api/notices?lastArticleId=${pageParam}&size=10&firstPage=false`);
+                const res = await axiosInstance.get(`/api/notices?lastArticleId=${pageParam}&size=10&firstPage=false`);
                 return res.data;
             }
             ,
@@ -46,12 +34,10 @@ const Notice: FC<{}> = () => {
         return data?.pages.flatMap((page) => page.result?.notices ?? []) ?? []; //2차원배열을 1차원배열로 변경해주는거
     }, [data]);
 
-    console.log(pages);
 
     React.useEffect(() => {
         if (inView) {
             fetchNextPage();
-            console.log(inView);
         }
     }, [inView]);
 
@@ -62,20 +48,7 @@ const Notice: FC<{}> = () => {
                 <span className="hidden md:inline-block md:text-slate-400 md:ml-4">
                     비즈큐레이터의 소식과 정보를 알려드립니다.
                 </span>
-                {/* {pages.map((i, index) => ( */}
                 <NoticeList dataList={pages} />
-                {/* ))} */}
-                {/* <button
-                    onClick={() => fetchPreviousPage()}
-                    disabled={!hasPreviousPage || isFetchingPreviousPage}
-                >
-                    {isFetchingPreviousPage
-                        ? "Loading more..."
-                        : hasPreviousPage
-                            ? "Load Older"
-                            : "Nothing more to load"}
-                </button> */}
-                {/* {isFetchingNextPage ? <p>Loading...</p> } */}
                 <div ref={ref}></div>
             </div>
         </Fragment>
