+++ components/customer/Question.tsx
@@ -1,9 +1,8 @@
-import React, { FC, Fragment } from "react";
+import React, { FC, Fragment, useMemo } from "react";
 import QuestionList from "./QuestionList";
-import { useQuery } from "@tanstack/react-query";
-import { atom, useRecoilValue } from "recoil";
-import axios from "axios";
-import { questionArticleIdState, questionizeState } from "@/atoms/questionAtom";
+import { useInfiniteQuery } from "@tanstack/react-query";
+import { useInView } from "react-intersection-observer";
+import axiosInstance from "@/apis/config";
 
 interface Item {
     id: number;
@@ -14,46 +13,44 @@ interface Item {
 
 const Question: FC<{}> = () => {
 
-    const questionArticleId = useRecoilValue(questionArticleIdState);
-    const questionSize = useRecoilValue(questionizeState);
+    const { ref, inView } = useInView();
 
-    const axiosData = async (questionArticleId: number, questionSize: number) => {
-        const path = "http://43.201.195.195:8080";
-        try {
-            const response = await axios.get<{
-                status: string;
-                code: number;
-                message: string;
-                result: {
-                    notices: Item[];
-                }
-            }>(
-                `${path}/api/faqs?lastArticleId=${questionArticleId}&size=${questionSize}`
-            );
-            return response.data.result.notices;
-        } catch (error) {
-            throw new Error("fail data");
+    const { data, fetchNextPage } = useInfiniteQuery<Item[], Error>(
+        ["question"],
+        async ({ pageParam = 0 }) => {
+            if (!pageParam) {
+                const res = await axiosInstance.get(`/api/faqs?lastArticleId=9999&size=10&firstPage=true`);
+                return res.data;
+            }
+            const res = await axiosInstance.get(`/api/faqs?lastArticleId=${pageParam}&size=10&firstPage=false`);
+            return res.data;
+        },
+        {
+            getNextPageParam: (lastPage) => {
+                const notices = lastPage?.result?.faqs ?? [];
+                const lastArticleId = notices[notices.length - 1]?.id ?? undefined
+                return lastArticleId;
+            },
         }
-    }
-
-    const { isLoading, isError, data, error } = useQuery<Item[], Error>(
-        [questionArticleId, questionSize],
-        () => axiosData(questionArticleId, questionSize)
     );
+    const pages = useMemo(() => {
+        return data?.pages.flatMap((page) => page.result?.faqs ?? []) ?? []; //2차원배열을 1차원배열로 변경해주는거
+    }, [data]);
 
-    if (isLoading) {
-        return <div>Loading...</div>;
-    }
-    if (isError) {
-        return <div>Error: {error?.message}</div>
-    }
+
+    React.useEffect(() => {
+        if (inView) {
+            fetchNextPage();
+        }
+    }, [inView]);
 
     return (
         <Fragment>
             <div className="md:absolute md:right-0 md:w-[800px]  md:mt-[88px]">
                 <h3 className="hidden md:inline-block md:text-2xl ">자주하는 질문</h3>
                 <span className="hidden md:inline-block md:text-slate-400 md:ml-4">고객님들께서 자주 해주시는 질문에 대한 응답입니다.</span>
-                <QuestionList item={data} />
+                <QuestionList item={pages} />
+                <div ref={ref}></div>
             </div>
         </Fragment>
     )
