+++ next.config.js
@@ -5,37 +5,7 @@ const nextConfig = {
     remotePatterns: [
       {
         protocol: 'https',
-        hostname: 'picsum.photos',
-        port: '',
-        pathname: '**',
-      },
-      {
-        protocol: 'https',
-        hostname: 'www.apple.com',
-        port: '',
-        pathname: '**',
-      },
-      {
-        protocol: 'https',
-        hostname: 'www.lge.co.kr',
-        port: '',
-        pathname: '**',
-      },
-      {
-        protocol: 'https',
-        hostname: 'gdimg.gmarket.co.kr',
-        port: '',
-        pathname: '**',
-      },
-      {
-        protocol: 'https',
-        hostname: 'm.ylshop.co.kr',
-        port: '',
-        pathname: '**',
-      },
-      {
-        protocol: 'https',
-        hostname: 'sitem.ssgcdn.com',
+        hostname: 's3-ebnatural.s3.ap-northeast-2.amazonaws.com',
         port: '',
         pathname: '**',
       },
