+++ pages/admin/VendorList.tsx
@@ -1,6 +1,5 @@
 import SidebarItem from "@/components/admin/SidebarItem";
 import AdminHeader from "@/components/admin/AdminHeader";
-import AdminSearch from "@/components/admin/AdminSearch";
 import AdminVendorList from "@/components/admin/VendorList";
 
 export default function VendorList() {
@@ -9,7 +8,6 @@ export default function VendorList() {
             <SidebarItem />
             <div className="ml-80 w-[1600px]">
                 <AdminHeader />
-                <AdminSearch />
                 <AdminVendorList />
             </div>
         </div>
