+++ utils/types/responseType.ts
@@ -1,5 +1,5 @@
 export interface NoticePostType {
-    itemId: number;
+    id: number;
     title: string;
     content: string;
     date: string;
